Refactor management tests to work with new layout
