+++ auth0/v2/test/management/test_device_credentials.py
@@ -1,11 +1,11 @@
 import unittest
 import mock
-from ..device_credentials import DeviceCredentials
+from ...management.device_credentials import DeviceCredentials
 
 
 class TestDeviceCredentials(unittest.TestCase):
 
-    @mock.patch('auth0.v2.device_credentials.RestClient')
+    @mock.patch('auth0.v2.management.device_credentials.RestClient')
     def test_get(self, mock_rc):
         mock_instance = mock_rc.return_value
 
@@ -21,7 +21,7 @@ class TestDeviceCredentials(unittest.TestCase):
                                             'client_id': 'cid',
                                             'type': 'type'})
 
-    @mock.patch('auth0.v2.device_credentials.RestClient')
+    @mock.patch('auth0.v2.management.device_credentials.RestClient')
     def test_create(self, mock_rc):
         mock_instance = mock_rc.return_value
 
@@ -33,7 +33,7 @@ class TestDeviceCredentials(unittest.TestCase):
         self.assertEqual('https://domain/api/v2/device-credentials', args[0])
         self.assertEqual(kwargs['data'], {'a': 'b', 'c': 'd'})
 
-    @mock.patch('auth0.v2.device_credentials.RestClient')
+    @mock.patch('auth0.v2.management.device_credentials.RestClient')
     def test_delete(self, mock_rc):
         mock_instance = mock_rc.return_value
