+++ auth0/v2/test/management/test_emails.py
@@ -1,11 +1,11 @@
 import unittest
 import mock
-from ..emails import Emails
+from ...management.emails import Emails
 
 
 class TestEmails(unittest.TestCase):
 
-    @mock.patch('auth0.v2.emails.RestClient')
+    @mock.patch('auth0.v2.management.emails.RestClient')
     def test_get(self, mock_rc):
         mock_instance = mock_rc.return_value
 
@@ -26,7 +26,7 @@ class TestEmails(unittest.TestCase):
         self.assertEqual(kwargs['params'], {'fields': 'a,b',
                                             'include_fields': 'false'})
 
-    @mock.patch('auth0.v2.emails.RestClient')
+    @mock.patch('auth0.v2.management.emails.RestClient')
     def test_config(self, mock_rc):
         mock_instance = mock_rc.return_value
 
@@ -38,7 +38,7 @@ class TestEmails(unittest.TestCase):
         self.assertEqual('https://domain/api/v2/emails/provider', args[0])
         self.assertEqual(kwargs['data'], {'a': 'b', 'c': 'd'})
 
-    @mock.patch('auth0.v2.emails.RestClient')
+    @mock.patch('auth0.v2.management.emails.RestClient')
     def test_update(self, mock_rc):
         mock_instance = mock_rc.return_value
 
@@ -50,7 +50,7 @@ class TestEmails(unittest.TestCase):
         self.assertEqual('https://domain/api/v2/emails/provider', args[0])
         self.assertEqual(kwargs['data'], {'a': 'b', 'c': 'd'})
 
-    @mock.patch('auth0.v2.emails.RestClient')
+    @mock.patch('auth0.v2.management.emails.RestClient')
     def test_delete(self, mock_rc):
         mock_instance = mock_rc.return_value
